+++ Runtime/UI/EmailRegisterDialogPresenter.cs
@@ -29,21 +29,18 @@ namespace com.noctuagames.sdk.UI
 
         //Behaviour whitelabel - VN
         private TextField _fullname;
-        private DropdownField _phoneCode;
         private TextField _phoneNumber;
         private TextField _birthDate;
         private DropdownField _gender;
-        private DropdownField _country;
+        private TextField _country;
         private TextField _idCard;
         private TextField _placeOfIssue;
-        private TextField _dateOfIssue;
+        private Button _dateOfIssue;
         private TextField _address;
         private List<TextField> textFields;
         private Button continueButton;
         private Button wizardContinueButton;
         private GlobalConfig _config;
-        private List<string> _phoneCodeList = new List<string>();
-        private List<string> _countryList = new List<string>();
 
         private int _wizardPage = 0;
 
@@ -96,6 +93,8 @@ namespace com.noctuagames.sdk.UI
             SetupInputFields(clearForm);
             HideAllErrors();
 
+            View.Q<VisualElement>("WizardSpinner").AddToClassList("hide");
+
             var additionalFooter = View.Q<VisualElement>("AdditionalFooterContent");
             if (isRegisterOnly)
             {
@@ -104,6 +103,19 @@ namespace com.noctuagames.sdk.UI
             else
             { 
                 additionalFooter.RemoveFromClassList("hide");
+            }
+
+            // Show copublisher logo
+            if (!string.IsNullOrEmpty(_config?.CoPublisher?.CompanyName))
+            {
+                var logo = Utility.GetCoPublisherLogo(_config.CoPublisher.CompanyName);
+                var defaultLogo = Resources.Load<Texture2D>(logo);
+                View.Q<VisualElement>("NoctuaLogoWithText").style.backgroundImage = new StyleBackground(defaultLogo);
+                View.Q<VisualElement>("NoctuaLogoWithText").RemoveFromClassList("hide");
+            }
+            else
+            {
+                View.Q<VisualElement>("NoctuaLogoWithText").AddToClassList("hide");
             }
 
             Visible = true;
@@ -160,22 +172,21 @@ namespace com.noctuagames.sdk.UI
 
             //Behaviour whitelabel - VN
             _fullname = View.Q<TextField>("FullNameTF");
-            _phoneCode = View.Q<DropdownField>("PhoneCodeDF");
             _phoneNumber = View.Q<TextField>("PhoneNumberTF");
             _birthDate = View.Q<TextField>("BirthdateTF");
             _gender = View.Q<DropdownField>("GenderTF");
-            _country = View.Q<DropdownField>("CountryTF");
+            _country = View.Q<TextField>("CountryTF");
             _idCard = View.Q<TextField>("IDCardTF");
             _placeOfIssue = View.Q<TextField>("PlaceOfIssueTF");
-            _dateOfIssue = View.Q<TextField>("DateOfIssueTF");
+            _dateOfIssue = View.Q<Button>("DateOfIssueTF");
             _address = View.Q<TextField>("AddressTF");
 
             if (!string.IsNullOrEmpty(_config?.Noctua?.Flags) && _config!.Noctua!.Flags!.Contains("VNLegalPurpose"))
             {
                 SetupDropdown();  
                 SetupDatePicker();
-                SetCountries();
-                ShowBehaviourWhitelabel(true);
+                // Disable country dropdown as Vietnam copublisher is asking for raw text input for country.
+                // SetCountries();
             }
 
             // Visibility
@@ -192,6 +203,7 @@ namespace com.noctuagames.sdk.UI
 
                 _fullname.value = "";
                 _phoneNumber.value = "";
+                _country.value = "";
                 _idCard.value = "";
                 _placeOfIssue.value = "";
                 _address.value = "";
@@ -207,9 +219,9 @@ namespace com.noctuagames.sdk.UI
                     _fullname,
                     _phoneNumber,
                     _birthDate,
+                    _country,
                     _idCard,
                     _placeOfIssue,
-                    _dateOfIssue,
                     _address
                 };
             }
@@ -257,6 +269,7 @@ namespace com.noctuagames.sdk.UI
             {
                 _fullname.RegisterValueChangedCallback(evt => OnValueChanged(_fullname));
                 _phoneNumber.RegisterValueChangedCallback(evt => OnValueChanged(_phoneNumber));
+                _country.RegisterValueChangedCallback(evt => OnValueChanged(_country));
                 _idCard.RegisterValueChangedCallback(evt => OnValueChanged(_idCard));
                 _placeOfIssue.RegisterValueChangedCallback(evt => OnValueChanged(_placeOfIssue));
                 _address.RegisterValueChangedCallback(evt => OnValueChanged(_address));
@@ -383,7 +396,7 @@ namespace com.noctuagames.sdk.UI
 
         private void SetupDropdown()
         {
-             var genderChoices = new List<string> {"Male", "Female"};
+             var genderChoices = new List<string> {Locale.GetTranslation("Select.Gender.Male"), Locale.GetTranslation("Select.Gender.Female")};
             Color textColor = new Color(98f / 255f, 100f / 255f, 104f / 255f);
             
             var regionCode = _config?.Noctua?.Region ?? "";
@@ -404,7 +417,6 @@ namespace com.noctuagames.sdk.UI
         private void SetupDatePicker()
         {
             _birthDate.isReadOnly = true;
-            _dateOfIssue.isReadOnly = true;
 
             string startDate = "01/01/2000";
             DateTime parsedDate = DateTime.ParseExact(startDate, "dd/MM/yyyy", null);
@@ -424,22 +436,58 @@ namespace com.noctuagames.sdk.UI
                     Debug.Log("Date Picked :" + _date.ToString("dd/MM/yyyy"));
 
                     textField.value = _date.ToString("dd/MM/yyyy");
-                    textField.labelElement.style.display = DisplayStyle.None;
-                    Utility.UpdateButtonState(textFields, continueButton);
-                    Utility.UpdateButtonState(textFields, wizardContinueButton);
                 });
             }
 
-            _birthDate.UnregisterCallback<PointerUpEvent>(evt => OpenDatePickerHandler(1, _birthDate, evt));
-            _dateOfIssue.UnregisterCallback<PointerUpEvent>(evt => OpenDatePickerHandler(2, _dateOfIssue, evt));
+            View.Q<VisualElement>("BirthdateContainer").UnregisterCallback<PointerUpEvent>(evt => {});
 
-            _birthDate.RegisterCallback<PointerUpEvent>(evt => OpenDatePickerHandler(1, _birthDate, evt));
-            _dateOfIssue.RegisterCallback<PointerUpEvent>(evt => OpenDatePickerHandler(2, _dateOfIssue, evt));  
+            View.Q<VisualElement>("BirthdateContainer").RegisterCallback<ClickEvent>(upEvent =>
+            {
+
+                upEvent.StopImmediatePropagation();
+
+                Noctua.OpenDatePicker(parsedDate.Year, parsedDate.Month, parsedDate.Day, 1,
+                (DateTime _date) =>
+                {
+                    _log.Debug($"picked date '{_date:O}'");
+                },
+                (DateTime _date) =>
+                {
+                    _birthDate.value = _date.ToString("dd/MM/yyyy");
+                    _birthDate.labelElement.style.display = DisplayStyle.None;
+                    Utility.UpdateButtonState(textFields, continueButton);
+                    Utility.UpdateButtonState(textFields, wizardContinueButton);
+
+                });
+            });
+
+            View.Q<VisualElement>("DateOfIssueContainer").UnregisterCallback<PointerUpEvent>(evt => {});
+            View.Q<VisualElement>("DateOfIssueContainer").RegisterCallback<ClickEvent>(upEvent =>
+            {
+
+                upEvent.StopImmediatePropagation();
+
+                Noctua.OpenDatePicker(parsedDate.Year, parsedDate.Month, parsedDate.Day, 1,
+                (DateTime _date) =>
+                {
+                    _log.Debug($"picked date '{_date:O}'");
+                },
+                (DateTime _date) =>
+                {
+                    _dateOfIssue.text = _date.ToString("dd/MM/yyyy");
+                    _dateOfIssue.RemoveFromClassList("grey-text");
+                    _dateOfIssue.AddToClassList("white-text");
+                    Utility.UpdateButtonState(textFields, continueButton);
+                    Utility.UpdateButtonState(textFields, wizardContinueButton);
+
+                });
+            });
         }
 
+        // Unused, leave it here for future use.
+        /*
         private void SetCountries()
         {
-            _phoneCodeList.Clear();
             _countryList.Clear();
 
             List<Country> countries = CountryData.Countries;
@@ -447,16 +495,7 @@ namespace com.noctuagames.sdk.UI
             foreach(var country in countries)
             {
                 _countryList.Add(country.Name);
-                _phoneCodeList.Add(country.PhoneCode);
             }
-
-            _phoneCode.choices = _phoneCodeList; 
-            _phoneCode.value = _phoneCodeList[0];
-            _phoneCode.RegisterCallback<ChangeEvent<string>>((evt) =>
-            {
-                _phoneCode.value = evt.newValue;
-                _phoneCode.labelElement.style.display = DisplayStyle.None;
-            });
 
             Color textColor = new Color(98f / 255f, 100f / 255f, 104f / 255f);
 
@@ -472,10 +511,13 @@ namespace com.noctuagames.sdk.UI
                 _country.labelElement.style.display = DisplayStyle.None;
             });            
         }
+        */
 
         private void OnLoginLinkClick(PointerUpEvent evt)
         {
             Visible = false;
+            
+            Model.ClearNavigation();
             Model.PushNavigation(() => Model.ShowEmailRegistration(false));
             Model.ShowEmailLogin();
         }
@@ -517,6 +559,13 @@ namespace com.noctuagames.sdk.UI
                 emailField.Q<Label>("error").RemoveFromClassList("hide");
                 emailField.Q<Label>("error").text = "Email address should not be empty";
                 emailField.Q<VisualElement>("title").style.color = ColorModule.redError;
+
+                // Show the error at the end of the wizard as well
+                if (!string.IsNullOrEmpty(_config?.Noctua?.Flags) && _config!.Noctua!.Flags!.Contains("VNLegalPurpose"))
+                {
+                    Model.ShowGeneralNotification(emailField.Q<Label>("error").text, false);
+                }
+
                 return;
             }
 
@@ -534,6 +583,13 @@ namespace com.noctuagames.sdk.UI
                 emailField.Q<Label>("error").RemoveFromClassList("hide");
                 emailField.Q<Label>("error").text = "Email address is not valid";
                 emailField.Q<VisualElement>("title").style.color = ColorModule.redError;
+
+                // Show the error at the end of the wizard as well
+                if (!string.IsNullOrEmpty(_config?.Noctua?.Flags) && _config!.Noctua!.Flags!.Contains("VNLegalPurpose"))
+                {
+                    Model.ShowGeneralNotification(emailField.Q<Label>("error").text, false);
+                }
+
                 return;
             }
 
@@ -551,6 +607,12 @@ namespace com.noctuagames.sdk.UI
                 passwordField.Q<Label>("error").RemoveFromClassList("hide");
                 passwordField.Q<Label>("error").text = "Password should not be empty";
                 passwordField.Q<VisualElement>("title").style.color = ColorModule.redError;
+
+                // Show the error at the end of the wizard as well
+                if (!string.IsNullOrEmpty(_config?.Noctua?.Flags) && _config!.Noctua!.Flags!.Contains("VNLegalPurpose"))
+                {
+                    Model.ShowGeneralNotification(passwordField.Q<Label>("error").text, false);
+                }
                 return;
             }
 
@@ -568,6 +630,12 @@ namespace com.noctuagames.sdk.UI
                 passwordField.Q<Label>("error").RemoveFromClassList("hide");
                 passwordField.Q<Label>("error").text = "Password is too short. Minimum 6 character";
                 passwordField.Q<VisualElement>("title").style.color = ColorModule.redError;
+
+                // Show the error at the end of the wizard as well
+                if (!string.IsNullOrEmpty(_config?.Noctua?.Flags) && _config!.Noctua!.Flags!.Contains("VNLegalPurpose"))
+                {
+                    Model.ShowGeneralNotification(passwordField.Q<Label>("error").text, false);
+                }
                 return;
             }
 
@@ -585,6 +653,12 @@ namespace com.noctuagames.sdk.UI
                 rePasswordField.Q<Label>("error").RemoveFromClassList("hide");
                 rePasswordField.Q<Label>("error").text = "Password is not matched with repeated password";
                 rePasswordField.Q<VisualElement>("title").style.color = ColorModule.redError;
+
+                // Show the error at the end of the wizard as well
+                if (!string.IsNullOrEmpty(_config?.Noctua?.Flags) && _config!.Noctua!.Flags!.Contains("VNLegalPurpose"))
+                {
+                    Model.ShowGeneralNotification(rePasswordField.Q<Label>("error").text, false);
+                }
                 return;
             }
 
@@ -608,25 +682,13 @@ namespace com.noctuagames.sdk.UI
                     _gender.Q<Label>("error").RemoveFromClassList("hide");
                     _gender.Q<Label>("error").text = "Please Select Gender!";
                     _gender.Q<VisualElement>("title").style.color = ColorModule.redError;
-                    return;
-                }
 
-                if(_country.value == "Select Country")
-                {
-                    _log.Debug("form validation: country is empty");
-                    //View.Q<Label>("ErrEmailEmpty").text = "Please Select Country!";
-                    //View.Q<Label>("ErrEmailEmpty").RemoveFromClassList("hide");
-                    View.Q<Button>("ContinueButton").RemoveFromClassList("hide");
-                    View.Q<VisualElement>("Spinner").AddToClassList("hide");
-                    // Wizard
-                    View.Q<Button>("WizardContinueButton").RemoveFromClassList("hide");            
-                    View.Q<Button>("WizardPrevTo3Button").RemoveFromClassList("hide");
-                    View.Q<VisualElement>("WizardSpinner").AddToClassList("hide");
+                    // Show the error at the end of the wizard as well
+                    if (!string.IsNullOrEmpty(_config?.Noctua?.Flags) && _config!.Noctua!.Flags!.Contains("VNLegalPurpose"))
+                    {
+                        Model.ShowGeneralNotification(_gender.Q<Label>("error").text, false);
+                    }
 
-                    _country.ElementAt(1).AddToClassList("noctua-text-input-error");
-                    _country.Q<Label>("error").RemoveFromClassList("hide");
-                    _country.Q<Label>("error").text = "Please Select Country!";
-                    _country.Q<VisualElement>("title").style.color = ColorModule.redError;
                     return;
                 }
 
@@ -650,18 +712,31 @@ namespace com.noctuagames.sdk.UI
                     _birthDate.Q<Label>("error").text = "Minimum age is 18 years old";
                     _birthDate.Q<VisualElement>("title").style.color = ColorModule.redError;
 
+                    // Show the error at the end of the wizard as well
+                    if (!string.IsNullOrEmpty(_config?.Noctua?.Flags) && _config!.Noctua!.Flags!.Contains("VNLegalPurpose"))
+                    {
+                        Model.ShowGeneralNotification(_birthDate.Q<Label>("error").text, false);
+                    }
+
+                    return;
+                }
+
+                if (_dateOfIssue.text == "")
+                {
+                    Model.ShowGeneralNotification("Date of issue should not be empty.", false);
                     return;
                 }
 
                 var issueDate = DateTime
-                    .ParseExact(_dateOfIssue.value, "dd/MM/yyyy", CultureInfo.InvariantCulture)
+                    .ParseExact(_dateOfIssue.text, "dd/MM/yyyy", CultureInfo.InvariantCulture)
                     .ToUniversalTime();
 
                 regExtra = new Dictionary<string, string>()
                 {
                     { "fullname", _fullname.value },
-                    { "phone_number", _phoneCode.value + _phoneNumber.value },
+                    { "phone_number", _phoneNumber.value },
                     { "birth_date", birthDate.ToString() },
+                    { "country", _country.value },
                     { "id_card", _idCard.value },
                     { "place_of_issue", _placeOfIssue.value },
                     { "date_of_issue", issueDate.ToString() },
@@ -682,11 +757,11 @@ namespace com.noctuagames.sdk.UI
                 if (!string.IsNullOrEmpty(_config?.Noctua?.Flags) && _config!.Noctua!.Flags!.Contains("VNLegalPurpose"))
                 {
                     _fullname.value = string.Empty;
-                    _phoneCode.value = string.Empty;
                     _birthDate.value = string.Empty;
+                    _country.value = string.Empty;
                     _idCard.value = string.Empty;
                     _placeOfIssue.value = string.Empty;
-                    _dateOfIssue.value = string.Empty;
+                    _dateOfIssue.text = string.Empty;
                     _address.value = string.Empty;
                 }
 
@@ -714,18 +789,24 @@ namespace com.noctuagames.sdk.UI
 
                 View.Q<Label>("ErrCode").RemoveFromClassList("hide");
                 View.Q<Button>("ContinueButton").RemoveFromClassList("hide");
-                View.Q<VisualElement>("AdditionalFooterContent").RemoveFromClassList("hide");
                 View.Q<VisualElement>("Spinner").AddToClassList("hide");
                 // Wizard
                 View.Q<Button>("WizardContinueButton").RemoveFromClassList("hide");            
                 View.Q<Button>("WizardPrevTo3Button").RemoveFromClassList("hide");
                 View.Q<VisualElement>("WizardSpinner").AddToClassList("hide");
+                if (!string.IsNullOrEmpty(_config?.Noctua?.Flags) && _config!.Noctua!.Flags!.Contains("VNLegalPurpose"))
+                {
+                    Model.ShowGeneralNotification(View.Q<Label>("ErrCode").text, false);
+                } else {
+                    View.Q<VisualElement>("AdditionalFooterContent").RemoveFromClassList("hide");
+                }
             }
         }
 
         private void OnBackButtonClick(PointerUpEvent evt)
         {
             _log.Debug("clicking back button");
+            View.Q<VisualElement>("WizardSpinner").RemoveFromClassList("hide");
 
             if (_wizardPage == 4) {
                 NavigateToWizard3();
@@ -825,36 +906,6 @@ namespace com.noctuagames.sdk.UI
             View.Q<Label>("ErrPasswordEmpty").AddToClassList("hide");
             View.Q<Label>("ErrPasswordMismatch").AddToClassList("hide");
             View.Q<Label>("ErrUnderage").AddToClassList("hide");
-        }
-
-        private void ShowBehaviourWhitelabel(bool isShow)
-        {
-            if(isShow)
-            {
-                View.Q<VisualElement>("PhoneNumberContainer").RemoveFromClassList("hide");
-
-                _fullname.RemoveFromClassList("hide");
-                _birthDate.RemoveFromClassList("hide");
-                _gender.RemoveFromClassList("hide");
-                _country.RemoveFromClassList("hide");
-                _idCard.RemoveFromClassList("hide");
-                _placeOfIssue.RemoveFromClassList("hide");
-                _dateOfIssue.RemoveFromClassList("hide");
-                _address.RemoveFromClassList("hide");
-            }
-            else
-            {
-                View.Q<VisualElement>("PhoneNumberContainer").AddToClassList("hide");
-
-                _fullname.AddToClassList("hide");
-                _birthDate.AddToClassList("hide");
-                _gender.AddToClassList("hide");
-                _country.AddToClassList("hide");
-                _idCard.AddToClassList("hide");
-                _placeOfIssue.AddToClassList("hide");
-                _dateOfIssue.AddToClassList("hide");
-                _address.AddToClassList("hide");
-            }
         }
     }
 }