+++ Runtime/Utility.cs
@@ -133,6 +133,23 @@ namespace com.noctuagames.sdk
                         {
                             var currentElement = root.Q<BindableElement>(name);
                             var currentValue = (currentElement as INotifyValueChanged<T>).value;
+                            
+                            if (currentValue == null && initialValues[name] == null)
+                            {
+                                continue;
+                            }
+                            
+                            if (currentValue == null && initialValues[name] != null)
+                            {
+                                anyChanged = true;
+                                break;
+                            }
+                            
+                            if (currentValue != null && initialValues[name] == null)
+                            {
+                                anyChanged = true;
+                                break;
+                            }
 
                             if (!currentValue.Equals(initialValues[name]))
                             {
@@ -140,7 +157,7 @@ namespace com.noctuagames.sdk
                                 break;
                             }
                         }
-                        buttonToEnable.SetEnabled(anyChanged);
+                        UpdateButtonState(buttonToEnable, anyChanged);                        
                     });
                 }
                 else
@@ -151,11 +168,14 @@ namespace com.noctuagames.sdk
         }
 
         public static void UpdateButtonState(List<TextField> textFields, Button submitButton)
-        {
-            bool isAnyFieldEmpty = textFields.Any(textField => string.IsNullOrEmpty(textField.value));
+        {           
+            UpdateButtonState(submitButton, !textFields.Any(textField => string.IsNullOrEmpty(textField.value)));
+        }
 
-            submitButton.SetEnabled(!isAnyFieldEmpty);
-            submitButton.pickingMode = !isAnyFieldEmpty ? PickingMode.Position : PickingMode.Ignore;
+        public static void UpdateButtonState(Button _submitButton, bool _isActive)
+        {
+            _submitButton.SetEnabled(_isActive);
+            _submitButton.pickingMode = _isActive ? PickingMode.Position : PickingMode.Ignore;
         }
 
         public static string GetCoPublisherLogo(string companyName)
@@ -253,9 +273,19 @@ namespace com.noctuagames.sdk
                     string textFieldTranslation = GetTranslation(textFieldKey, translations);
                     textField.label = textFieldTranslation;
 
-                    Label labelTitle = textField.Q<Label>("title");
+                    Label textFieldTitle = textField.Q<Label>("title");
 
-                    if (labelTitle != null) labelTitle.text = textFieldTranslation;                    
+                    if (textFieldTitle != null) textFieldTitle.text = textFieldTranslation;                    
+
+                    break;
+                case DropdownField dropdownField:
+                    string dropdownFieldKey = $"{uxmlName}.{elementName}.{elementType}.label";
+                    string dropdownFieldTranslation = GetTranslation(dropdownFieldKey, translations);
+                    dropdownField.label = dropdownFieldTranslation;
+
+                    Label dropdownTitle = dropdownField.Q<Label>("title");
+
+                    if (dropdownTitle != null) dropdownTitle.text = dropdownFieldTranslation;                    
 
                     break;
                 case VisualElement visualElement: